requirements.txt
langchain
python-dotenv
langchain_groq
langchain_google_genai
langchain_community
pypdf
## langchain.document_loaders
faiss-cpu
structlog
PyMuPDF
-e .

langchain_openai
langchain_core
sys

